+++ src/pieces/rook.hpp
@@ -6,9 +6,9 @@
 
 #include "utility/binary.hpp"
 
-std::uint64_t get_rook_xrayed_squares_from_mailbox(std::uint8_t mb) noexcept;
-std::uint64_t get_rook_blocker_squares_from_mailbox(std::uint8_t mb) noexcept;
-std::uint64_t get_rook_attacked_squares_from_mailbox(std::uint64_t pos, std::uint8_t mb) noexcept;
+std::uint64_t get_rook_xrayed_squares_from_mailbox(std::size_t mb) noexcept;
+std::uint64_t get_rook_blocker_squares_from_mailbox(std::size_t mb) noexcept;
+std::uint64_t get_rook_attacked_squares_from_mailbox(std::uint64_t pos, std::size_t mb) noexcept;
 
 inline std::uint64_t get_rook_attacked_squares_from_mailboxes(std::uint64_t pos, auto... mbs) noexcept { return (get_rook_attacked_squares_from_mailbox(pos, mbs) | ...); }
 
@@ -23,21 +23,21 @@ inline std::uint64_t get_rook_attacked_squares_from_mailboxes(std::uint64_t pos,
 namespace details
 {
 
-inline std::uint64_t get_rook_xrayed_squares_from_mailbox_impl(std::uint8_t mb) noexcept
+inline std::uint64_t get_rook_xrayed_squares_from_mailbox_impl(std::size_t mb) noexcept
 {
     return (get_bitboard_mailbox_rank(mb) | get_bitboard_mailbox_file(mb)) & ~get_bitboard_mailbox_piece(mb);
 }
 
 extern const std::span<std::uint64_t> xray_table_rook;
 
-inline std::uint64_t get_rook_blocker_squares_from_mailbox_impl(std::uint8_t mb) noexcept
+inline std::uint64_t get_rook_blocker_squares_from_mailbox_impl(std::size_t mb) noexcept
 {
     return ((get_bitboard_mailbox_rank(mb) & ~FILE_A & ~FILE_H) | (get_bitboard_mailbox_file(mb) & ~RANK_1 & ~RANK_8)) & ~get_bitboard_mailbox_piece(mb);
 }
 
 extern const std::span<std::uint64_t> blocker_table_rook;
 
-inline std::uint64_t get_rook_attacked_squares_from_mailbox_impl(std::uint8_t mb, std::uint64_t pos)
+inline std::uint64_t get_rook_attacked_squares_from_mailbox_impl(std::size_t mb, std::uint64_t pos)
 {
     std::uint64_t ret {};
 
@@ -62,17 +62,17 @@ extern const std::array<details::pext_bitboard, 64>  attack_table_rook;
 
 }
 
-inline std::uint64_t get_rook_xrayed_squares_from_mailbox(std::uint8_t mb) noexcept
+inline std::uint64_t get_rook_xrayed_squares_from_mailbox(std::size_t mb) noexcept
 {
     return details::xray_table_rook[mb];
 }
 
-inline std::uint64_t get_rook_blocker_squares_from_mailbox(std::uint8_t mb) noexcept
+inline std::uint64_t get_rook_blocker_squares_from_mailbox(std::size_t mb) noexcept
 {
     return details::blocker_table_rook[mb];
 }
 
-inline std::uint64_t get_rook_attacked_squares_from_mailbox(std::uint64_t pos, std::uint8_t mb) noexcept
+inline std::uint64_t get_rook_attacked_squares_from_mailbox(std::uint64_t pos, std::size_t mb) noexcept
 {
     return details::attack_table_rook[mb][pos];
 }
