+++ src/utility/coordinates.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstdint>
 #include <string>
 
-std::string to_coordinates_str(std::uint8_t mb);
-std::uint8_t from_coordinates_str(std::string_view str);
+std::string to_coordinates_str(std::size_t mb);
+std::size_t from_coordinates_str(std::string_view str);
